11 refactor to serial (#12)
* refactor

* updating to serial id

+++ board-game-persistance/src/main/java/com/coding2themax/boardgame/persistance/repo/BoardGameRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
 import com.coding2themax.boardgame.persistance.model.BoardGame;
 
-public interface BoardGameRepository extends ReactiveCrudRepository<BoardGame, Integer> {
+public interface BoardGameRepository extends ReactiveCrudRepository<BoardGame, Long> {
 }
